+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Possibly fix a crash that would occur when dispatching analytics events - alloy
 * The fairs rail is now constrained to just it's tab - orta
 * Re-enable Lots by Artists You Follow and hide it when there are no live sale lots - alloy
+* Don’t crash on trying to JSON parse a network error - alloy
 
 ### 1.4.1